Remove more unused or commented code
