+++ ffc/uflacs/analysis/graph.py
@@ -13,12 +13,12 @@ from ffc.uflacs.analysis.graph_symbols import build_graph_symbols
 class Graph2(object):
     def __init__(self):
 
+        # Index to expression and expression to index maps
         self.V = []
         self.e2i = {}
 
         self.expression_vertices = []
 
-        self.V_shapes = []
         self.V_symbols = None
         self.total_unique_symbols = 0
 
@@ -31,6 +31,6 @@ def build_graph(expressions):
     G.e2i, G.V, G.expression_vertices = build_graph_vertices(expressions, scalar=False)
 
     # Populate with symbols
-    G.V_shapes, G.V_symbols, G.total_unique_symbols = build_graph_symbols(G.V)
+    G.V_symbols, G.total_unique_symbols = build_graph_symbols(G.V)
 
     return G
